Update the systemd libs used to build docker image
The libraries created in the docker image can not read the journal db in newer
versions of systemd. This specifically impacts users that run systemd `246` or
above on their host server:

> * systemd-journald gained support for zstd compression of large fields in
> journal files. The hash tables in journal files have been hardened against
> hash collisions. This is an incompatible change and means that journal files
> created with new systemd versions are not readable with old versions. If the
> $SYSTEMD_JOURNAL_KEYED_HASH boolean environment variable for
> systemd-journald.service is set to 0 this new hardening functionality may be
> turned off, so that generated journal files remain compatible with older
> journalctl implementations.

(Taken from https://github.com/systemd/systemd/blob/v246/NEWS)

This commit uses the backported version of the systemd libs, which are `247` for
Debian buster. Fixing it for most distributions.

It might be worth considering having a separate build stage for systemd, that
would give better flexibility of the version shipped in the image.

Signed-off-by: Yasser Saleemi <yassersaleemi@gmail.com>